// src/common/enums/http-status.enum.ts
export enum HttpStatus {
    OK = 200,
    CREATED = 201,
    BAD_REQUEST = 400,
    UNAUTHORIZED = 401,
    FORBIDDEN = 403,
    NOT_FOUND = 404,
    INTERNAL_SERVER_ERROR = 500,

    CONTINUE = 100,  
    ACCEPTED = 202,
    NON_AUTHORITATIVE_INFORMATION = 203,   
    MOVED_PERMANENTLY = 301,
    FOUND = 302,
    NOT_MODIFIED = 304,
    PAYMENT_REQUIRED = 402,
    TOO_MANY_REQUESTS = 429,
    NOT_IMPLEMENTED = 501,
    BAD_GATEWAY = 502,
    SERVICE_UNAVAILABLE = 503,
    GATEWAY_TIMEOUT = 504,

}